+++ src/storage/storage.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 import { StorageService } from './storage.service';
-import { Backup, BackupSchema } from 'src/storage/schemas/storage.schema';
+import { BackupLMData, BackupSchema } from 'src/storage/schemas/storage.schema';
 
 @Module({
   imports: [
-    MongooseModule.forFeature([{ name: Backup.name, schema: BackupSchema }]),
+    MongooseModule.forFeature([
+      { name: BackupLMData.name, schema: BackupSchema },
+    ]),
   ],
   providers: [StorageService],
 })
