+++ src/tools/tools.module.ts
@@ -5,10 +5,12 @@ import { BackupService } from './backup.service';
 import { UtilsService } from '../utils/utils.service';
 import { StorageService } from '../storage/storage.service';
 import { MongooseModule } from '@nestjs/mongoose';
-import { Backup, BackupSchema } from '../storage/schemas/storage.schema';
+import { BackupLMData, BackupSchema } from '../storage/schemas/storage.schema';
 @Module({
   imports: [
-    MongooseModule.forFeature([{ name: Backup.name, schema: BackupSchema }]),
+    MongooseModule.forFeature([
+      { name: BackupLMData.name, schema: BackupSchema },
+    ]),
   ],
   controllers: [ToolsController],
   providers: [ToolsService, BackupService, UtilsService, StorageService],
